Extract and rename MarketHoursDatabase.Key class for reuse
